gRPC server enhancements (#296)
* Identify connections to gRPC server with Uuids

* Add concurrent request and subscription handlers to the connection to the gRPC server

* On shutdown clear services from core and async runtime

* Use tonic v0.10.x

* Make `Notifier` sync channel available to integration tests

* Fix tokio test runtime in daemon integration tests

* Remove links from Connection & ConnectionHandler to Manager

* Stop gRPC server forcefully on shutdown timeout

* Add a gRPC client-server sanity check test

* Lint

* Reduce sleep time

* Refactor gRPC client-server test

* Add a gRPC client-server notifications test

* Allocate only one thread to tokio test runtime of all daemon integration tests

* Handle legacy virtual hash in `EstimateNetworkHashesPerSecond` RPC method

* Make `subscribe` fn resistant to a theoretical race-condition

* Ignore duplicated subscriptions

* Add notify requests to Scope conversions

* Unify subscription handler logic (one exception)

* Fix meta

* Fix typos

* Add a gRPC payload ops enum

* Replace `RpcApiOps` with `KaspadPayloadOps` in gRPC client

* Replace `RpcApiOps` with `KaspadPayloadOps` in gRPC server

* Add a macro building a gRPC server interface

* Give request handlers payload op specific methods

* Fix doc

* Working GetMetrics method in gRPC client & server

* Remove obsolete `RpcApiOps::Notification`

* Remove `KaspadPayloadOps::is_subscription()`

* Change GetMetrics server time type to u64

* Strictly correlate interface method set to KaspadPayloadOps variant set

* Set gRPC server shutdown timeout to 1 second

* Update tonic, prost & h2 to the latest versions

* Move the connection routing logic inside the request receive loop, saving a lock

* Make gRPC logs more consistent

* Disconnect the client if the outgoing route gets full

* Ignore oneshot send potential error

* Polish logs

* Refactor "not implemented" error building

+++ rpc/core/src/model/message.rs
@@ -718,13 +718,13 @@ pub struct ConsensusMetrics {
 #[derive(Clone, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, BorshSchema)]
 #[serde(rename_all = "camelCase")]
 pub struct GetMetricsResponse {
-    pub server_time: u128,
+    pub server_time: u64,
     pub process_metrics: Option<ProcessMetrics>,
     pub consensus_metrics: Option<ConsensusMetrics>,
 }
 
 impl GetMetricsResponse {
-    pub fn new(server_time: u128, process_metrics: Option<ProcessMetrics>, consensus_metrics: Option<ConsensusMetrics>) -> Self {
+    pub fn new(server_time: u64, process_metrics: Option<ProcessMetrics>, consensus_metrics: Option<ConsensusMetrics>) -> Self {
         Self { process_metrics, consensus_metrics, server_time }
     }
 }
@@ -1068,7 +1068,7 @@ impl SubscribeResponse {
 }
 
 ///
-///  wRPC response for RpcApiOps::Subscribe request
+///  wRPC response for RpcApiOps::Unsubscribe request
 ///
 #[derive(Clone, Debug, Serialize, Deserialize, BorshSerialize, BorshDeserialize, BorshSchema)]
 #[serde(rename_all = "camelCase")]
